+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Release Notes
 
+### 0.1.0
+- Export Stored Params to CSV/TSV file
+- Import Params from CSV/TSV file
+- UI Update
+
 ### 0.0.2
 - Bug Fixes
 - New Feature "Replace By Group"